feat: add base types
