+++ libs/@local/hash-graph-types/graph_types/_schema.py
@@ -1,14 +1,18 @@
 import asyncio
 from abc import ABC, abstractmethod
 from collections.abc import Awaitable
+from types import EllipsisType
 from typing import TYPE_CHECKING, Annotated, Any, Generic, Literal, TypeVar
 from uuid import UUID
 
 from pydantic import (
     BaseModel,
+    ConfigDict,
     Field,
     GetCoreSchemaHandler,
     GetJsonSchemaHandler,
+    conlist,
+    create_model,
 )
 from pydantic.json_schema import JsonSchemaValue
 from pydantic_core import CoreSchema, core_schema
@@ -91,7 +95,7 @@ class Array(Schema, Generic[T]):
         *,
         actor_id: UUID,
         graph: "GraphAPIProtocol",
-    ) -> Annotated[Any, ...]:
+    ) -> type[list[T]]:
         ty = await self.items.create_model(actor_id=actor_id, graph=graph)
 
         class ListSchema:
@@ -107,7 +111,7 @@ class Array(Schema, Generic[T]):
                     max_length=self.max_items,
                 )
 
-        return Annotated[list[T], ListSchema]
+        return conlist(ty, min_length=self.min_items, max_length=self.max_items)
 
 
 class Object(Schema, Generic[T]):
@@ -127,6 +131,12 @@ class Object(Schema, Generic[T]):
         ) -> tuple[str, type[BaseModel] | Any]:
             return key, await value
 
+        def default_value(key: str) -> None | EllipsisType:
+            if key in self.required:
+                return ...
+
+            return None
+
         types = dict(
             await asyncio.gather(
                 *(
@@ -136,36 +146,10 @@ class Object(Schema, Generic[T]):
             ),
         )
 
-        class DictSchema:
-            @classmethod
-            def __get_pydantic_core_schema__(
-                cls,
-                _source_type: Any,  # noqa: ANN401
-                handler: GetCoreSchemaHandler,
-            ) -> CoreSchema:
-                return core_schema.typed_dict_schema(
-                    {
-                        property_type_id: core_schema.typed_dict_field(
-                            handler.generate_schema(property_type),
-                            required=(
-                                property_type_id in self.required
-                                if self.required
-                                else None
-                            ),
-                        )
-                        for property_type_id, property_type in types.items()
-                    },
-                    extra_behavior="forbid",
-                )
+        types = {key: (value, default_value(key)) for key, value in types.items()}
 
-            @classmethod
-            def __get_pydantic_json_schema__(
-                cls,
-                schema: CoreSchema,
-                handler: GetJsonSchemaHandler,
-            ) -> JsonSchemaValue:
-                json_schema = handler(schema)
-                json_schema.update(additionalProperties=False)
-                return json_schema
-
-        return Annotated[dict[str, Any], DictSchema]
+        return create_model(
+            "DictSchema",
+            __config__=ConfigDict(extra="forbid"),
+            **types,
+        )
